Reverting work on executing chrome for final streams CL
