+++ chrome_agent/background.html
@@ -3,7 +3,7 @@
 	<head>
 		<script src="js/lib/axiom_base.amd.concat.js"></script>
 		<script src="js/lib/wash.amd.concat.js"></script>
-		<script src="js/chrome_command.js"></script>
+		<!-- TODO (ericarnold): implement: script src="js/chrome_command.js"></script-->
 		<script src="js/chrome_agent.js"></script>
 		<script src="js/extension/background.js"></script>
 	</head>
