+++ lib/wash/exe/chrome.js
