+++ lib/wash/exe_modules.js
@@ -3,37 +3,39 @@ export var dir = {};
 export default dir;
 import m0 from "wash/exe/cat";
 dir["cat"] = m0;
-import m1 from "wash/exe/clear";
-dir["clear"] = m1;
-import m2 from "wash/exe/cp";
-dir["cp"] = m2;
-import m3 from "wash/exe/echo";
-dir["echo"] = m3;
-import m4 from "wash/exe/eval";
-dir["eval"] = m4;
-import m5 from "wash/exe/import";
-dir["import"] = m5;
-import m6 from "wash/exe/ls";
-dir["ls"] = m6;
-import m7 from "wash/exe/mkdir";
-dir["mkdir"] = m7;
-import m8 from "wash/exe/mkstream";
-dir["mkstream"] = m8;
-import m9 from "wash/exe/mount.gdrive";
-dir["mount.gdrive"] = m9;
-import m10 from "wash/exe/mount";
-dir["mount"] = m10;
-import m11 from "wash/exe/mount.stream";
-dir["mount.stream"] = m11;
-import m12 from "wash/exe/mv";
-dir["mv"] = m12;
-import m13 from "wash/exe/pwd";
-dir["pwd"] = m13;
-import m14 from "wash/exe/readline";
-dir["readline"] = m14;
-import m15 from "wash/exe/rm";
-dir["rm"] = m15;
-import m16 from "wash/exe/touch";
-dir["touch"] = m16;
-import m17 from "wash/exe/wash";
-dir["wash"] = m17;
+import m1 from "wash/exe/chrome";
+dir["chrome"] = m1;
+import m2 from "wash/exe/clear";
+dir["clear"] = m2;
+import m3 from "wash/exe/cp";
+dir["cp"] = m3;
+import m4 from "wash/exe/echo";
+dir["echo"] = m4;
+import m5 from "wash/exe/eval";
+dir["eval"] = m5;
+import m6 from "wash/exe/import";
+dir["import"] = m6;
+import m7 from "wash/exe/ls";
+dir["ls"] = m7;
+import m8 from "wash/exe/mkdir";
+dir["mkdir"] = m8;
+import m9 from "wash/exe/mkstream";
+dir["mkstream"] = m9;
+import m10 from "wash/exe/mount.gdrive";
+dir["mount.gdrive"] = m10;
+import m11 from "wash/exe/mount";
+dir["mount"] = m11;
+import m12 from "wash/exe/mount.stream";
+dir["mount.stream"] = m12;
+import m13 from "wash/exe/mv";
+dir["mv"] = m13;
+import m14 from "wash/exe/pwd";
+dir["pwd"] = m14;
+import m15 from "wash/exe/readline";
+dir["readline"] = m15;
+import m16 from "wash/exe/rm";
+dir["rm"] = m16;
+import m17 from "wash/exe/touch";
+dir["touch"] = m17;
+import m18 from "wash/exe/wash";
+dir["wash"] = m18;
