+++ lib/wash/wash_util.js
@@ -32,7 +32,6 @@ var StatResult;
 washUtil.findExecutable = function(cx, pathSpec) {
   var rootPath = cx.fileSystemManager.defaultFileSystem.rootPath;
   var searchList = cx.getEnv('@PATH', [rootPath.spec]);
-  var localPath = Path.abs(cx.getPwd(), pathSpec);
 
   /** @type {function(): !Promise<!{path: !Path, statResult: StatResult}>} */
   var searchNextPath = function() {
@@ -60,21 +59,5 @@ washUtil.findExecutable = function(cx, pathSpec) {
     });
   };
 
-  return cx.fileSystemManager.stat(localPath).then(
-    function(statResult) {
-      if (statResult.mode & Path.Mode.X) {
-        return Promise.resolve({
-          fileSystem: this.fileSystemManager,
-          path: localPath,
-          statResult: statResult
-        });
-      }
-      return searchNextPath();
-    }.bind(this)
-  ).catch(function(value) {
-    return searchNextPath();
-  }.bind(this));
+  return searchNextPath();
 };
-
-
-
